chore(modal): apply sb guidelines to modal
